+++ src/dataclass_extensions/__init__.py
@@ -3,13 +3,11 @@ from .encode import encode
 from .merge import merge, merge_from_dotlist
 from .registrable import Registrable
 from .types import Dataclass
-from .utils import required_field
 
 __all__ = [
     "Dataclass",
     "Registrable",
     "DecodeError",
-    "required_field",
     "encode",
     "decode",
     "merge",