+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `strict` argument to `encode()` is now deprecated. Prefer `errors` instead.
 
+### Removed
+
+- Removed `required_field`.
+
 ## [v0.4.0](https://github.com/epwalsh/dataclass-extensions/releases/tag/v0.4.0) - 2026-03-05
 
 ### Added